스프링 카페 2주차 3단계 구현 (#659)
* Remove: remove unused InMemoryRepository class

* Implement mission 2-3: add database schema of reply

* Implement mission 2-3: add reply domain

* Implement mission 2-3: add reply feature (update not included)

* Implement mission 2-3: add Dockerfile

* Implement mission 2-3: migrate from H2 to MySQL

* Implement mission 2-3: prepare deployment using docker-compose

+++ src/main/java/com/kakao/cafe/service/ArticleService.java
@@ -1,10 +1,12 @@
 package com.kakao.cafe.service;
 
 import com.kakao.cafe.domain.article.Article;
+import com.kakao.cafe.domain.reply.Reply;
 import com.kakao.cafe.domain.user.User;
 import com.kakao.cafe.exception.NoSuchArticleException;
 import com.kakao.cafe.exception.UnauthenticatedArticleAccessException;
 import com.kakao.cafe.repository.ArticleRepository;
+import com.kakao.cafe.repository.ReplyRepository;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.stereotype.Service;
@@ -13,9 +15,11 @@ import org.springframework.stereotype.Service;
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
+    private final ReplyRepository replyRepository;
 
-    public ArticleService(ArticleRepository articleRepository) {
+    public ArticleService(ArticleRepository articleRepository, ReplyRepository replyRepository) {
         this.articleRepository = articleRepository;
+        this.replyRepository = replyRepository;
     }
 
     public void registerArticle(Article article) {
@@ -58,6 +62,13 @@ public class ArticleService {
         if (!article.isWriter(user)) {
             throw new UnauthenticatedArticleAccessException("다른 사람의 게시글을 수정, 삭제할 수 없습니다");
         }
+        List<Reply> replyList = replyRepository.findAllByArticleId(articleId);
+        boolean isArticleDeletable = replyList.stream().
+                allMatch((reply -> reply.isWriter(user)));
+        if (!isArticleDeletable) {
+            throw new UnauthenticatedArticleAccessException("다른 사람의 댓글이 존재할 경우 삭제할 수 없습니다.");
+        }
+        replyRepository.deleteByArticleId(articleId);
         articleRepository.delete(article);
     }
 }
